Try to reduce apc warning related to smarty. Closes T3553
Summary: We use APC for opcache. We seem to get too many warnings deleting the item from cache, we should avoid deleting the file from cache too often. So we add some naive protection - if the file on the disk is less then 15 seconds old, we do not try to write it on the disk again. Chances are that the file is already the same on the disk and we would only overwrite it with same content.

Test Plan:
Terminals A and B:

1) @A: tail -f /var/log/apache2/error.log
2) @B: vg-purge-tpl
3) Open http://vagrant.dev:8080/apc.php?OB=2&SCOPE=A&SORT1=H&SORT2=D&COUNT=0&SEARCH=layout.tpl
3a) Load http://beta.shoptet.cz.dev:8080/
4) touch /vagrant-src/templates/frontend_templates/05/templates/layout/layout.tpl
4a) Load http://beta.shoptet.cz.dev:8080/
5) Open in new TAB http://vagrant.dev:8080/apc.php?OB=2&SCOPE=A&SORT1=H&SORT2=D&COUNT=0&SEARCH=layout.tpl
6) Compare 3 and 5 - "Created at" timestamp must change

Reviewers: fercak

Reviewed By: fercak

Maniphest Tasks: T3553

Differential Revision: https://phabricator.shoptet.cz:8443/D2158